Reorganize support contexts
* Collapse into single support context -- "with holdings tables" -- that
  truncates all tables

* Remove most helpers for adding data (since data types can use create / save
  now); leave htitem since that is done externally (hathifiles_database)
