+++ spec/integration_spec.rb
@@ -8,9 +8,7 @@ RSpec.describe "phctl integration" do
     PHCTL::PHCTL.start(args)
   end
 
-  include_context "with cluster ocns table"
-  include_context "with hathifiles table"
-  include_context "with holdings table"
+  include_context "with tables for holdings"
 
   describe "load" do
     xdescribe "commitments" do
