+++ spec/spec_helper.rb
@@ -16,9 +16,7 @@ require "services"
 require "sidekiq/batch"
 require "rspec-sidekiq"
 require "fileutils"
-require_relative "support/cluster_ocn_table"
-require_relative "support/hathifiles_table"
-require_relative "support/holdings_table"
+require_relative "support/holdings_tables"
 
 SimpleCov::Formatter::LcovFormatter.config do |c|
   c.report_with_single_file = true
